A general rename, campaign should be called an experiment.
