+++ docs/src/modules/testing/nav.adoc
@@ -6,10 +6,10 @@
 *** xref:testing:evaluation/concepts.adoc[Concepts and vocabulary]
 *** xref:testing:evaluation/eval-cases-and-rules.adoc[Eval cases and rules]
 *** xref:testing:evaluation/evaluators.adoc[Evaluators]
-*** xref:testing:evaluation/campaigns.adoc[Campaigns and runs]
+*** xref:testing:evaluation/experiments.adoc[Experiments and runs]
 *** xref:testing:evaluation/reports.adoc[Reports]
 *** xref:testing:evaluation/baselines.adoc[Baselines and regression]
-*** xref:testing:evaluation/remote-campaigns.adoc[Remote campaigns]
+*** xref:testing:evaluation/remote-experiments.adoc[Remote experiments]
 ** xref:testing:red-teaming/index.adoc[Red teaming]
 *** xref:testing:red-teaming/getting-started.adoc[Getting started]
 *** xref:testing:red-teaming/rules-and-hazards.adoc[Rules and hazards]
@@ -17,6 +17,6 @@
 *** xref:testing:red-teaming/techniques.adoc[Attack styles]
 *** xref:testing:red-teaming/adaptive-attacks.adoc[Adaptive attacks]
 *** xref:testing:red-teaming/corpora.adoc[Corpora]
-*** xref:testing:red-teaming/campaigns.adoc[Campaigns and runs]
+*** xref:testing:red-teaming/experiments.adoc[Experiments and runs]
 *** xref:testing:red-teaming/reports.adoc[Reports]
 *** xref:testing:red-teaming/attestation.adoc[Signed evidence]
