+++ docs/src/modules/testing/pages/evaluation/index.adoc
@@ -1,7 +1,7 @@
 = Evaluation
 :page-summary: An evaluation suite runs a service against a set of eval cases and records a verdict for each. Reports carry pass rates, the cases that failed, and the cost of the run. A CI job can gate a merge on regression.
 :page-when-to-use: Use when a service is worth measuring. The reply's correctness, the tool usage, the latency budget, the faithfulness of a retrieved answer.
-:page-related: xref:testing:evaluation/getting-started.adoc[], xref:testing:evaluation/concepts.adoc[], xref:testing:evaluation/evaluators.adoc[], xref:testing:evaluation/eval-cases-and-rules.adoc[], xref:testing:evaluation/campaigns.adoc[], xref:testing:evaluation/reports.adoc[], xref:testing:red-teaming/index.adoc[]
+:page-related: xref:testing:evaluation/getting-started.adoc[], xref:testing:evaluation/concepts.adoc[], xref:testing:evaluation/evaluators.adoc[], xref:testing:evaluation/eval-cases-and-rules.adoc[], xref:testing:evaluation/experiments.adoc[], xref:testing:evaluation/reports.adoc[], xref:testing:red-teaming/index.adoc[]
 :page-persona: developer
 :page-banner: evaluations
 
@@ -12,7 +12,7 @@ A suite runs the service through a fixed set of eval cases and turns each reply 
 The rendered report lists the failing cases one by one.
 The CI job fails the build when the pass rate drops or a case that used to pass now fails.
 
-== The parts of a campaign
+== The parts of an experiment
 
 `EvalCase` +
 A graded turn, what a correct answer does, an optional specification node or metric, and an
@@ -28,17 +28,17 @@ See xref:testing:evaluation/evaluators.adoc[].
 
 `EvalCaseResults` +
 One eval case together with every attempt made against it.
-Repeating a campaign produces several attempts and one `EvalCaseResults`.
+Repeating an experiment produces several attempts and one `EvalCaseResults`.
 The report reads results case by case.
 An eval case that passed twice and failed once is one varied verdict.
 
-`Campaign` +
+`Experiment` +
 The JUnit class that assembles the eval cases, the target and the evaluators into a named
 unit and runs them.
 Gated on `-Deval=true` so `mvn verify` compiles the class without running it.
 
 `Run` +
-One execution of a campaign against one target build.
+One execution of an experiment against one target build.
 
 `Report` +
 The output of a run.
@@ -53,7 +53,7 @@ A test adapter, `SystemUnderTest`, puts the Agent in front of the runner.
 
 == See also
 
-* xref:testing:evaluation/getting-started.adoc[]. From an empty project through the first campaign.
+* xref:testing:evaluation/getting-started.adoc[]. From an empty project through the first experiment.
 * xref:testing:evaluation/evaluators.adoc[]. The three families, the built-ins, and how to add your own.
 * xref:testing:evaluation/reports.adoc[]. Every panel of the rendered report.
 * xref:testing:red-teaming/index.adoc[]. Adversarial suite that reuses the same target adapter.
