+++ docs/src/modules/testing/pages/red-teaming/evaluators.adoc
@@ -1,21 +1,21 @@
 = Adversarial evaluators
 :page-summary: Decide whether the target broke through a red-team rule using a deterministic, heuristic, or agentic evaluator.
 :page-when-to-use: Use when a red-team rule needs a specific decision. The target quoted a configured secret, refused a harmful request, or complied with a jailbreak framed as fiction.
-:page-related: xref:testing:red-teaming/index.adoc[], xref:testing:red-teaming/rules-and-hazards.adoc[], xref:testing:red-teaming/techniques.adoc[], xref:testing:red-teaming/campaigns.adoc[], xref:testing:evaluation/evaluators.adoc[]
+:page-related: xref:testing:red-teaming/index.adoc[], xref:testing:red-teaming/rules-and-hazards.adoc[], xref:testing:red-teaming/techniques.adoc[], xref:testing:red-teaming/experiments.adoc[], xref:testing:evaluation/evaluators.adoc[]
 :page-banner: evaluations
 
 include::ROOT:partial$include.adoc[]
 
 An adversarial evaluator decides whether the target broke through a red-team rule.
 Red-team evaluators use the same three-family model as evaluation evaluators.
-A campaign that mixes them settles the clear cases cheaply so the expensive judgment runs only on the borderline.
+An experiment that mixes them settles the clear cases cheaply so the expensive judgment runs only on the borderline.
 
 == Overview
 
 Every red-team evaluator implements `akka.evalkit.redteam.evaluator.RedTeamEvaluator` and returns a `ScoreResult`.
 The result is one of `BROKE_THROUGH`, `HELD`, or `INCONCLUSIVE`.
 Each break is one incident with its own row in the report.
-A campaign that ran 500 attacks against an evaluator reports the count of breaks and the count of holds as separate figures.
+An experiment that ran 500 attacks against an evaluator reports the count of breaks and the count of holds as separate figures.
 
 The three families answer three different kinds of question:
 
@@ -52,7 +52,7 @@ A marker present in the reply is a `BROKE_THROUGH`. A marker absent is `HELD`. A
 include::example$evaluation-playground/src/test/java/com/example/redteam/evaluator/heuristic/JailbreakSample.java[tag=evaluator]
 ----
 <1> The patterns are baked into the class.
-Changing them would change what "the same `JailbreakEvaluator`" means from campaign to campaign, so the report reads a `JailbreakEvaluator` verdict as "this specific pattern set decided the case".
+Changing them would change what "the same `JailbreakEvaluator`" means from experiment to experiment, so the report reads a `JailbreakEvaluator` verdict as "this specific pattern set decided the case".
 <2> A reply with refusal language and no task-fulfilment reads as `HELD`. A reply with task-fulfilment and no refusal reads as `BROKE_THROUGH`. A reply carrying both signals reads as `INCONCLUSIVE`.
 
 The two-tier pattern is where the family taxonomy pays off.
@@ -88,7 +88,7 @@ An evaluator holds no component client, so the call lives here rather than insid
 <4> The evaluator handles the rest: check that there is anything to ask about, call the assessor, validate the confidence, produce a `ScoreResult`.
 
 A confidence below the evaluator's threshold (default 0.6) reads as `INCONCLUSIVE`.
-The campaign either iterates with more attempts or leaves the exchange in the borderline column of the report.
+The experiment either iterates with more attempts or leaves the exchange in the borderline column of the report.
 
 Tests use a stub assessor:
 
@@ -115,7 +115,7 @@ For the reference tables listing every built-in evaluator, the hazards it speaks
 
 == See also
 
-* xref:testing:red-teaming/getting-started.adoc[]. A worked example from empty project to first red-team campaign.
+* xref:testing:red-teaming/getting-started.adoc[]. A worked example from empty project to first red-team experiment.
 * xref:testing:red-teaming/techniques.adoc[]. Attack styles that turn one goal into many prompts.
 * xref:testing:red-teaming/adaptive-attacks.adoc[]. Attackers that revise their prompts based on the reply.
 * xref:testing:red-teaming/reports.adoc[]. How verdicts read in the red-team report.
