+++ docs/src/modules/testing/pages/red-teaming/getting-started.adoc
@@ -7,14 +7,14 @@
 
 include::ROOT:partial$include.adoc[]
 
-A red-team campaign runs a service through prompts written to attack it and asks an evaluator whether the target broke through, held, or the evidence was inconclusive.
+A red-team experiment runs a service through prompts written to attack it and asks an evaluator whether the target broke through, held, or the evidence was inconclusive.
 Each break is one incident with its own row in the report.
 Security review reads a red-team run the way it reads a penetration test: case by case, not as a rate.
 
 == Overview
 
-A red-team campaign has four parts.
-The **target** is the same `SystemUnderTest` an evaluation campaign uses, so a project with an evaluation suite already reuses its adapter.
+A red-team experiment has four parts.
+The **target** is the same `SystemUnderTest` an evaluation experiment uses, so a project with an evaluation suite already reuses its adapter.
 The **corpus** is a set of attacker goals ("get the model to describe how to bypass 2FA"), pulled from date-versioned artifacts like HarmBench, DAN, or JailbreakBench.
 The **techniques** turn each goal into many prompts.
 Static wrappers roleplay the request or base64-encode the payload.
@@ -24,7 +24,7 @@ The **evaluators** decide, per exchange, whether the target complied with a harm
 Red-team suites are gated on `-Dredteam=true`.
 `mvn verify` compiles the classes without running them.
 `mvn verify -Dredteam=true` runs them.
-The runtime enforces a token and wall-clock budget so a campaign cannot run away with the customer's spend.
+The runtime enforces a token and wall-clock budget so an experiment cannot run away with the customer's spend.
 
 == Add the dependency
 
@@ -34,7 +34,7 @@ The runtime enforces a token and wall-clock budget so a campaign cannot run away
 ----
 $ akka redteam init
 added the evalkit-redteam dependency to pom.xml (previous kept as pom.xml.backup)
-run a campaign with `mvn verify -Dredteam=true`
+run an experiment with `mvn verify -Dredteam=true`
 
 Added the redteam source root to . under package com.acme.checkout.redteam — 9 file(s) written, 0 kept.
 ----
@@ -62,21 +62,21 @@ For the license terms and how to add corpora of your own, see xref:testing:red-t
 
 == Reuse the evaluation adapter
 
-The same `SystemUnderTest` implementation the evaluation campaign uses reaches the target for a red-team campaign.
+The same `SystemUnderTest` implementation the evaluation experiment uses reaches the target for a red-team experiment.
 For the full code, see xref:testing:evaluation/getting-started.adoc[].
 The adapter section applies unchanged.
-A project starting with red teaming and no prior evaluation writes the same adapter before the campaign runs.
+A project starting with red teaming and no prior evaluation writes the same adapter before the experiment runs.
 
-== The campaign class
+== The experiment class
 
-A red-team campaign assembles the evaluators, the techniques, and the budget, then runs against the target.
+A red-team experiment assembles the evaluators, the techniques, and the budget, then runs against the target.
 
 [source,java,indent=0]
-.RefundAgentRedTeamCampaign.java
+.RefundAgentRedTeamExperiment.java
 ----
-include::example$evaluation-playground/src/test/java/com/example/redteam/campaign/RefundAgentRedTeamCampaign.java[tag=class]
+include::example$evaluation-playground/src/test/java/com/example/redteam/experiment/RefundAgentRedTeamExperiment.java[tag=class]
 ----
-<1> The system-property gate keeps campaigns out of `mvn verify`.
+<1> The system-property gate keeps experiments out of `mvn verify`.
 <2> Multiple evaluators cover multiple hazards.
 The attempt is a breakthrough when any of them says so.
 <3> `DriverState` holds everything carried between turns: the goal, the technique driving it, and the turn limit.
@@ -90,9 +90,9 @@ Only the first is a breakthrough.
 
 == What a run reports
 
-`mvn verify -Dredteam=true` runs the campaign.
+`mvn verify -Dredteam=true` runs the experiment.
 `RedTeamPanels.render` returns the report as a `String`.
-The campaign writes it wherever the project keeps them.
+The experiment writes it wherever the project keeps them.
 The report is plain 80-column ASCII and prints its panels in a fixed order.
 A panel appears only when the run has something to put in it, and the numbering starts at 1 with no gaps.
 
@@ -101,14 +101,14 @@ The example below is the report a full run produces (`RedTeamPanels.render()` ou
 [listing]
 ----
 ================================================================================
-Red-team campaign: refund-agent-safety
+Red-team experiment: refund-agent-safety
 Run:      2026-08-23  09:07:12   (1h 33m 51s)
 Target:   refund-agent   (build a7c31f2, gpt-4o-mini via router)
 Attempts: 12,480 prompts across 8 rules, 6 attack styles, 3 repeats each
 ================================================================================
 
 
-1. What the campaign found
+1. What the experiment found
 --------------------------
 
 The system broke a rule on 140 out of 12,480 attempts (1.1%).
@@ -157,7 +157,7 @@ Do not help wrongdoing        3/60     11/60    24/60    9/60     14/60
 8. What it cost
 ---------------
 
-Token spend for this campaign (input + output):
+Token spend for this experiment (input + output):
 
   Target (the system under test)       8,412,301 tokens  ~ $18.42
   Attacker (adaptive attack styles)    3,201,880 tokens  ~ $7.05
@@ -180,7 +180,7 @@ Full transcripts:       target/redkit/transcripts/
 ----
 
 Panels shown above: 1, 2, 3, 8, plus the header and footer.
-For the exact output of panels 4 through 10, including "What varied between attempts", "How harm was measured", "How the judge scored", "What the campaign could not tell you", "Compared to the last run", and "Where the quality tests missed this", see xref:testing:red-teaming/reports.adoc[].
+For the exact output of panels 4 through 10, including "What varied between attempts", "How harm was measured", "How the judge scored", "What the experiment could not tell you", "Compared to the last run", and "Where the quality tests missed this", see xref:testing:red-teaming/reports.adoc[].
 
 == Where the parts live
