+++ docs/src/modules/testing/pages/red-teaming/rules-and-hazards.adoc
@@ -7,7 +7,7 @@
 include::ROOT:partial$include.adoc[]
 
 A rule under red teaming is the same object as a rule under evaluation.
-An evaluation campaign runs the rule under normal input to confirm it holds; a red-team campaign runs the rule under adversarial input to see whether it breaks.
+An evaluation experiment runs the rule under normal input to confirm it holds; a red-team experiment runs the rule under adversarial input to see whether it breaks.
 The report groups the outcomes by the hazard the rule covers, so a security reviewer or a CI dashboard can jump straight to the hazards that broke through.
 
 == Declaring a rule with hazards
